update:modal分包对mapper的影响 add:task运行完成保存消费记录

+++ celloud-v3/src/main/java/com/celloud/constants/PriceType.java
@@ -10,5 +10,5 @@ public class PriceType {
     /**
      * item_id为APP编号
      */
-    public static final int isApp = 0;
+    public static final Byte isApp = 0;
 }
